requirements.txt
pytest==8.3.5
pytest-mock==3.14.0
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.11.1
python-magic==0.4.27
pdfplumber==0.11.6
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
pytesseract==0.3.13
Pillow==11.1.0
scikit-learn==1.6.1
numpy==2.2.4
rapidfuzz==3.12.2